added dummy audio recommendations

+++ JusAudio/app/src/main/java/com/adrienshen_n_vlad/jus_audio/persistence/entities/PlayHistory.kt
@@ -1,10 +1,13 @@
 package com.adrienshen_n_vlad.jus_audio.persistence.entities
 
 import androidx.annotation.Keep
-import androidx.room.*
-import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.ROW_ID
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
 import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.AUDIO_STREAM_URL
 import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.HISTORY_TABLE_NAME
+import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.ROW_ID
 
 @Keep
 @Entity(
@@ -16,6 +19,6 @@ import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.HISTORY
 )
 data class PlayHistory(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = ROW_ID) var rowId: Int,
+    @ColumnInfo(name = ROW_ID) var rowId: Long,
     @ColumnInfo(name = AUDIO_STREAM_URL) var audioStreamUrl: String
 )
