+++ JusAudio/app/src/main/java/com/adrienshen_n_vlad/jus_audio/persistence/entities/RecommendedAudios.kt
@@ -1,7 +1,10 @@
 package com.adrienshen_n_vlad.jus_audio.persistence.entities
 
 import androidx.annotation.Keep
-import androidx.room.*
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.Index
+import androidx.room.PrimaryKey
 import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants
 import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.RECENTLY_RECOMMENDED_TABLE_NAME
 
@@ -16,6 +19,6 @@ import com.adrienshen_n_vlad.jus_audio.utility_classes.JusAudioConstants.RECENTL
 )
 data class RecommendedAudios(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = JusAudioConstants.ROW_ID) var rowId: Int,
+    @ColumnInfo(name = JusAudioConstants.ROW_ID) var rowId: Long,
     @ColumnInfo(name = JusAudioConstants.AUDIO_STREAM_URL) var audioStreamUrl: String
 )
